[FIX] global numbering on MacOSX in optimized mode
git-svn-id: https://coolfluidsrv.vki.ac.be/svn/coolfluid3/trunk@3121 4892e656-8fb9-448a-b5bc-b1919152596f

+++ src/Mesh/Actions/CGlobalNumbering.cpp
@@ -308,7 +308,7 @@ std::size_t CGlobalNumbering::hash_value(const RealVector& coords)
 {
   std::size_t seed=0;
   for (Uint i=0; i<coords.size(); ++i)
-    boost::hash_combine(seed,(float)coords[i]);
+    boost::hash_combine(seed,coords[i]);
   return seed;
 }
 
@@ -319,7 +319,7 @@ std::size_t CGlobalNumbering::hash_value(const RealMatrix& coords)
   std::size_t seed=0;
   for (Uint i=0; i<coords.rows(); ++i)
   for (Uint j=0; j<coords.cols(); ++j)
-    boost::hash_combine(seed,(float)coords(i,j));
+    boost::hash_combine(seed,coords(i,j));
   return seed;
 }
